feat: Refactor BM25 search engine into a separate module for improved maintainability and modularity
