+++ src/lib/web-mcp/modules/bm25/index.ts
@@ -1,0 +1,2 @@
+export { BM25SearchEngine } from './bm25-search-engine';
+export type { BM25SearchResult, LocalBM25Instance } from './types';
